requirements.txt
# Flask dan dependencies
Flask==2.3.3
Flask-CORS==4.0.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# YOLOv8
ultralytics==8.0.196

# Additional utilities
requests==2.31.0
python-dotenv==1.0.0

# Development
gunicorn==21.2.0 